Extract tests for isTypeError to own file
